Split provider test into separate files
